feat: Add to_json method to TextMessage class (#312)
* feat: Add to_json method to TextMessage class

Add a to_json method to the TextMessage class to convert its content to
JSON format. As such, the return type TextMessage can be used as part
of a multi-turn conversion and be used as part of the chat history.
A test that would have caught the missing serialization option was 
added.